requirements.txt
pygame-ce==2.5.7
numpy
matplotlib
sys
os
csv
time
subprocess
datetime
